Modifications to prevent "leakage" of function arguments (that should remain local variables within the function) into the calling script.  Details below.
When a script calls a function which in turn uses the process_args function to parse the function arguments, those "dummy" argument variables should only be defined within the scope of the function; they should not become defined in the scope of the calling script since they may conflict with variables of the same name within the calling script.

This is currently an issue when using process_args because proceess_args simply uses "eval" to evaluate the key-value pairs passed to it, where the key is the name of the argument variable and the value is the value it should take on.  This approach causes these new arguments/variables to be defined not only within the scope of process_args but also within the scope of the function that calls process_args (which we want) as well as any other scripts/functions that call the function (Which we don't want).

The solution is as follows.  Within process_args, instead of using "eval" to set the arguments/variables to the specified values, construct a string that contains the commands that need to be executed in order for those arguments/variables to be set to the desired values.  Then have process_args pass this string to the calling function (just by printing it to stdout), and have that function perform an "eval" on that string.  In addition, when process_args is called by a function (as opposed to a script), include in the command string the "local" modifiers, i.e. instead of using

some_argument="some_value"

use

local some_argument="some_value"

With this approach, the calling function must then use "eval" to execute that string containing the commands.  This will cause the arguments to be local variables within that function and not "leak out" to any other scripts/functions calling the function.

Note that when the entity that is calling process_args is a script (instead of a function), then we can't use the "local" modifier in the command string that process_args creates because using "eval" to execute that command string within the calling script will result in an error.  That is because in bash, local variables are not allowed within scripts; they are only allowed within functions.